+++ packages/elementary_tools/elementary_cli/lib/commands/COMMANDS.md
@@ -1,0 +1,9 @@
+
+## Structure
+
+Every folder represents a 'branch' command and contains file with same name.
+
+* branch_command_1
+* - branch_command_1.dart
+* - leaf_command_1.dart
+* - leaf_command_2.dart
