+++ packages/elementary_tools/elementary_cli/lib/commands/generate/generate.dart
@@ -5,7 +5,7 @@ import 'dart:isolate';
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:elementary_cli/commands/generate/generate_module.dart';
-import 'package:elementary_cli/commands/generate/generate_test.dart';
+import 'package:elementary_cli/commands/generate/generate_test/generate_test.dart';
 import 'package:elementary_cli/exit_code_exception.dart';
 import 'package:path/path.dart' as p;
