+++ packages/elementary_tools/elementary_cli/lib/commands/generate/generate_test/generate_test.dart
@@ -1,5 +1,5 @@
 import 'package:args/command_runner.dart';
-import 'package:elementary_cli/commands/generate/generate_test_wm.dart';
+import 'package:elementary_cli/commands/generate/generate_test/generate_test_wm.dart';
 
 /// `elementary_tools generate test` command
 class GenerateTestCommand extends Command<void> {
